+++ src/components/posts/Post.jsx
@@ -1,9 +1,29 @@
 import React from 'react'
+import styled from 'styled-components'
 
-import { Stack } from '../../../components/styles/Stack.styled'
-import { Tag } from './styles/Tag.styled'
-import { TopicBar } from './styles/TopicBar.styled'
-import { PostCard } from './styles/PostCard.styled'
+const PostCard = styled.div`
+  border: 1px solid rgba(30, 30, 100, 0.8);
+  border-radius: 3px;
+  margin-bottom: 20px;
+  width: 100%;
+`
+
+const Tag = styled.div`
+  font-size: 0.8rem;
+  background-color: rgba(30, 30, 30, 0.5);
+  color: #ccc;
+  padding: 3px 5px;
+  border-radius: 3px;
+  margin-top: 2px;
+  margin-right: 3px;
+  margin-bottom: 2px;
+`
+
+export const TopicBar = styled.div`
+  background-color: rgba(30, 30, 100, 0.8);
+  padding: 2px 0px 2px 4px;
+  display: flex;
+`
 
 const Post = ({ postData }) => {
   const { title, topics, post, postBody, timestamp } = postData
@@ -11,7 +31,7 @@ const Post = ({ postData }) => {
   const postedAt = timestamp?.toDate().toString() || null
 
   return (
-    <Stack gutter='xs'>
+    
       <PostCard>
         {topics ? (
           <TopicBar>
@@ -24,7 +44,7 @@ const Post = ({ postData }) => {
         <p>{post || postBody}</p>
         {postedAt && <p>Posted at: {postedAt}</p>}
       </PostCard>
-    </Stack>
+    
   )
 }
