--- src/components/styles/Wrapper.styled.js
@@ -1,17 +1,0 @@
-// Import styles components
-import styled from 'styled-components'
-
-// Global container that wraps the whole app
-export const GlobalContainer = styled.div`
-  @import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
-
-  font-family: 'Poppins', sans-serif;
-  padding: none;
-  margin: none;
-  box-sizing: border-box;
-
-  a {
-    text-decoration: none;
-    color: black;
-  }
-`
