--- src/features/README.md
@@ -1,3 +1,0 @@
-# FEATURES
-
-This folder contains feature based modules - i.e. pages or groups of pages related to a specific feature.
