--- src/features/posts/components/styles/PostCard.styled.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const PostCard = styled.div`
-  border: 1px solid rgba(30, 30, 100, 0.8);
-  border-radius: 3px;
-  margin-bottom: 5px;
-`
