--- src/features/posts/components/styles/Tag.styled.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-export const Tag = styled.div`
-  font-size: 0.8rem;
-  background-color: rgba(30, 30, 30, 0.5);
-  color: #ccc;
-  padding: 3px 5px;
-  border-radius: 3px;
-  margin-top: 2px;
-  margin-right: 3px;
-  margin-bottom: 2px;
-`
