--- src/features/posts/components/styles/TopicBar.styled.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const TopicBar = styled.div`
-  background-color: rgba(30, 30, 100, 0.8);
-  padding: 2px 0px 2px 4px;
-  display: flex;
-`
