+++ src/pages/Profile.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { useAuthContext } from '../../hooks/useAuthContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const Profile = () => {
   const { user } = useAuthContext()
